services: Hack backup eligibility checks to backup more application data
Seedvault is our backup transporter; it produces a local, encrypted backup
of application data (and optionally the application itself). This means we
no longer need to adhere to strictly following the allowAutoBackup flag as
we aren't backuping to an insecure medium like Google does with their GMS
backup utility.

Furthermore, avoiding backing up an application just because it was force
stopped is stupidity - what has that got to do with the application being
backed up or not?

We preserve the disabled application check to save space.

Signed-off-by: Diab Neiroukh <lazerl0rd@thezest.dev>